Memleak Fix
Using the d3dx12.h version of comparison for D3D12_RESOURCE_DESC, compared the Alignment member as well which is usually passed in as 0 for default selection by the app. This does not match the Alignment retrieved by calling GetDesc() on the resource. This was causing resourse reuse to fail and run out of memory.